separated tests into separate parts
